Make signature and transaction version uint32
