+++ wallet/crypto/ed25519.go
@@ -54,6 +54,6 @@ func (e *ed25519Sig) Name() string {
 	return "vega/ed25519"
 }
 
-func (e *ed25519Sig) Version() uint64 {
+func (e *ed25519Sig) Version() uint32 {
 	return 1
 }
