+++ wallet/handler_deprecated.go
@@ -9,28 +9,6 @@ import (
 	"github.com/golang/protobuf/proto"
 )
 
-func (h *Handler) SignAny(token, inputData, pubKey string) ([]byte, error) {
-	h.mu.RLock()
-	defer h.mu.RUnlock()
-
-	// first the transaction would be in base64, let's decode
-	rawInputData, err := base64.StdEncoding.DecodeString(inputData)
-	if err != nil {
-		return nil, err
-	}
-
-	kp, err := h.getKeyPair(token, pubKey)
-	if err != nil {
-		return nil, err
-	}
-
-	if kp.Tainted {
-		return nil, ErrPubKeyIsTainted
-	}
-
-	return kp.Algorithm.Sign(kp.privBytes, rawInputData)
-}
-
 func (h *Handler) SignTx(token, tx, pubKey string) (SignedBundle, error) {
 	h.mu.RLock()
 	defer h.mu.RUnlock()
